+++ packages/react/src/components/AppShell/src/contexts/AppShellMenuContext/AppShellMenuContext.definitions.ts
@@ -1,0 +1,5 @@
+import { AppShellMenuContextValue } from "./AppShellMenuContext.types";
+
+export const initialAppShellMenuContextValue: AppShellMenuContextValue = {
+  isMenuPopover: false,
+} as const;